refactor: split out mobility BFF requests into api/bff/
