+++ src/modules/mobility/queries/use-bike-station-query.tsx
@@ -1,5 +1,5 @@
 import {useQuery} from '@tanstack/react-query';
-import {getBikeStation} from '@atb/api/mobility';
+import {getBikeStation} from '@atb/api/bff/mobility';
 import {ONE_MINUTE_MS} from '@atb/utils/durations';
 
 export const useBikeStationQuery = (id: string) =>
