+++ src/modules/mobility/queries/use-vehicle-query.tsx
@@ -1,5 +1,5 @@
 import {useQuery} from '@tanstack/react-query';
-import {getVehicle} from '@atb/api/mobility';
+import {getVehicle} from '@atb/api/bff/mobility';
 import {ONE_MINUTE_MS} from '@atb/utils/durations';
 
 export const useVehicleQuery = (id: string) =>
