+++ src/modules/mobility/use-stations.tsx
@@ -12,7 +12,7 @@ import {
 import {Point} from 'geojson';
 import {StationBasicFragment} from '@atb/api/types/generated/fragments/stations';
 import {useIsFocused} from '@react-navigation/native';
-import {getStations} from '@atb/api/mobility';
+import {getStations} from '@atb/api/bff/mobility';
 import {FormFactor} from '@atb/api/types/generated/mobility-types_v2';
 import {useFeatureTogglesContext} from '@atb/modules/feature-toggles';
