+++ src/modules/parking-violations-reporting/use-parking-violations.tsx
@@ -1,6 +1,6 @@
 import {useEffect, useState} from 'react';
 import {useGeolocationContext} from '@atb/modules/geolocation';
-import {initViolationsReporting} from '@atb/api/mobility';
+import {initViolationsReporting} from '@atb/api/bff/mobility';
 import {
   ParkingViolationType,
   ViolationsReportingProvider,
